bdev/nvme: add entry_ctx_in_use
We will now keep a list of the possible paths to
the discovery subsystem.  One of them will be the
path we are currently connected to (which at service
start, is the path specified by the user).

Additional entries are added for discovery log page
entries referencing the discovery subsystem.

When the discovery service starts, we just have the
initial entry in the list - the discovery poller
tries to connect to it, and if the connect starts
successfully, removes it from the list and points
ctx->entry_ctx_in_use to it.

This will be useful later when we want to iterate
through the available paths to the discovery
subsystem if the current path fails.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I5b18e0f20c4607e29ac0f12f27ba7eb169d0206d
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/11770
Community-CI: Mellanox Build Bot
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@mellanox.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>